+++ src/cli/commands/call.js
@@ -42,7 +42,6 @@ exports.desecription = DESCRIPTION;
 exports.builder = (yargs) => {
   yargs
     .usage(USAGE)
-    .demand(1)
     .options(_.merge({
       data: {
         description: `Specify inline the JSON data to send to the function. ${'Default:'.bold} ${'{}'.green}`,
@@ -81,28 +80,7 @@ exports.handler = (opts) => {
   const controller = new Controller(opts);
 
   return controller.doIfRunning()
-    .then(() => {
-      if (controller.server.get('inspect')) {
-        if (controller.server.get('isolation') === 'inprocess') {
-          console.log(`Function execution paused. Connect to the debugger on port ${controller.server.get('inspectPort')} (e.g. using the "node" launch type in VSCode).\n`);
-        } else {
-          controller.getDebuggingUrl().then((debugUrl) => {
-            let debugStr = `Function execution paused. Connect to the debugger on port ${controller.server.get('inspectPort')} (e.g. using the "node2" launch type in VSCode), or open the following URL in Chrome:`;
-            if (debugUrl) {
-              // If found, include it in the string that gets printed
-              debugStr += `\n\n    ${debugUrl}\n`;
-            } else {
-              debugStr += `\n\nError: Could not find Chrome debugging URL in log file. Look for it yourself in ${controller.server.get('logFile')}.`;
-            }
-            console.log(debugStr);
-          });
-        }
-      } else if (controller.server.get('debug')) {
-        console.log(`Function execution paused. Connect to the debugger on port ${controller.server.get('debugPort')} (e.g. using the "node" launch type in VSCode).\n`);
-      }
-
-      return controller.call(opts.functionName, opts.data);
-    })
+    .then(() => controller.call(opts.functionName, opts.data))
     .then(([body, response]) => {
       controller.log(`ExecutionId: ${body.executionId}`);
       if (body.result) {